fltk::app

Function handle

pub fn handle<I: Into<Event> + Copy + PartialEq + PartialOrd, W: WindowExt>(
    msg: I,
    w: &W,
) -> Result<bool, FltkError>

Send a signal to a window. Integral values from 0 to 30 are reserved. Returns Ok(true) if the event was handled. Returns Ok(false) if the event was not handled. Returns Err on error or in use of one of the reserved values.

use fltk::{prelude::*, *};
const CHANGE_FRAME: i32 = 100;
let mut wind = window::Window::default();
let mut but = button::Button::default();
let mut frame = frame::Frame::default();
but.set_callback(move |_| {
    let _ = app::handle(CHANGE_FRAME, &wind).unwrap();
});
frame.handle(move |f, ev| {
    if ev == CHANGE_FRAME.into() {
        f.set_label("Hello world");
        true
    } else {
        false
    }
});

Errors

Returns Err on error or in use of one of the reserved values.