fltk::app

Function handle_raw

pub unsafe fn handle_raw(event: Event, w: WindowPtr) -> bool

Send a signal to a window pointer from event_dispatch. Returns true if the event was handled. Returns false if the event was not handled or ignored.

use fltk::{prelude::*, *};
const CHANGE_FRAME: i32 = 100;
let mut wind = window::Window::default();
let mut but = button::Button::default();
let mut frame = frame::Frame::default();
wind.end();
wind.show();
but.set_callback(move |_| {
    let _ = app::handle_main(CHANGE_FRAME).unwrap();
});

frame.handle(move |f, ev| {
    if ev == CHANGE_FRAME.into() {
        f.set_label("Hello world");
        true
    } else {
        false
    }
});
unsafe {
    app::event_dispatch(|ev, winptr| {
        if ev == CHANGE_FRAME.into() {
            false // ignore CHANGE_FRAME event
        } else {
            app::handle_raw(ev, winptr)
        }
    });
}

Safety

The window pointer must be valid