Module fltk::dialog

Dialog widgets

Structs

FileChooser

FLTK’s own FileChooser. Which differs for the Native FileDialog Example:

FileChooserType

The types of FileChooser

FileDialog

Creates a file button

HelpDialog

Creates a help dialog

Enums

BeepType

Defines the type of beep to be passed to the beep function

ColorMode

Color modes to be used with the color chooser

FileDialogOptions

Defines the File dialog options, which can be set using the set_option() method.

FileDialogType

Defines the type of dialog, which can be changed dynamically using the set_type() method

Functions

alert

Displays an alert box

alert_default

Displays an alert box

beep

Emits a beep

choice

Displays a choice box with upto three choices. An empty choice will not be shown

choice_default

Displays a choice box with upto three choices. An empty choice will not be shown

color_chooser

Spawns a color chooser dialog.

color_chooser_with_default

Spawns a color chooser dialog.

dir_chooser

Shows a directory chooser returning a String

file_chooser

Shows a file chooser returning a String. Example:

input

Displays an input box, which returns the inputted string. Can be used for gui io

input_default

Displays an input box, which returns the inputted string. Can be used for gui io

message

Displays a message box

message_default

Displays a message box

password

Shows an input box, but with hidden string

password_default

Shows an input box, but with hidden string

Type Definitions

NativeFileChooser

Re-alias FileDialog to NativeFileChooser (Fl_Native_File_Chooser)

NativeFileChooserOptions

Alias to NativeFileChooserOptions

NativeFileChooserType

Alias for NativeFileChooserType