Struct fltk::dialog::FileChooser

pub struct FileChooser { /* fields omitted */ }

FLTK’s own FileChooser. Which differs for the Native FileDialog Example:

use fltk::{prelude::*, *};
fn main() {
    let app = app::App::default();
    let mut win = window::Window::default().with_size(900, 300);
    let mut chooser = dialog::FileChooser::new(
        ".",                    // directory
        "*",                    // filter or pattern
        dialog::FileChooserType::Multi, // chooser type
        "Title Of Chooser",     // title
    );
    chooser.show();
    chooser.window().set_pos(300, 300);
    // Block until user picks something.
    //     (The other way to do this is to use a callback())
    //
    while chooser.shown() {
        app::wait();
    }
    // User hit cancel?
    if chooser.value(1).is_none() {
        println!("(User hit 'Cancel')");
        return;
    }
    // Print what the user picked
    println!("--------------------");
    println!("DIRECTORY: '{}'", chooser.directory().unwrap());
    println!("    VALUE: '{}'", chooser.value(1).unwrap()); // value starts at 1!
    println!("    COUNT: {} files selected", chooser.count());
    // Multiple files? Show all of them
    if chooser.count() > 1 {
        for t in 1..=chooser.count() {
            println!(" VALUE[{}]: '{}'", t, chooser.value(t).unwrap());
        }
    }
    win.end();
    win.show();
    app.run().unwrap();
}

Implementations

impl FileChooser

pub fn new(
dir: &str,
pattern: &str,
typ: FileChooserType,
title: &str
) -> FileChooser

Instantiates a new FileChooser

pub unsafe fn delete(dlg: Self)

Deletes a FileChooser

Safety

Can invalidate the underlying pointer

pub fn new_button(&mut self) -> Option<impl ButtonExt>

Gets the new button of the FileChooser

pub fn preview_button(&mut self) -> Option<impl ButtonExt>

Gets the preview button of the FileChooser

pub fn show_hidden_button(&mut self) -> Option<impl ButtonExt>

Gets the show hidden button of the FileChooser

pub fn set_callback<F: FnMut(&mut Self) + 'static>(&mut self, cb: F)

Sets the callback of the FileChooser

pub fn set_color(&mut self, c: Color)

Sets the color of the FileChooser

pub fn color(&mut self) -> Color

Gets the color of the FileChooser

pub fn count(&mut self) -> i32

Gets the count of chosen items

pub fn set_directory(&mut self, dir: &str)

Sets the directory of the FileChooser

pub fn directory(&mut self) -> Option<String>

Gets the directory of the FileChooser

pub fn set_filter(&mut self, pattern: &str)

Sets the filter for the dialog, can be: A single wildcard (eg. "*.txt"). Multiple wildcards (eg. "*.{cxx,h,H}"). A descriptive name followed by a \t and a wildcard (eg. "Text Files\t*.txt"). A list of separate wildcards with a \n between each (eg. "*.{cxx,H}\n*.txt"). A list of descriptive names and wildcards (eg. "C++ Files\t*.{cxx,H}\nTxt Files\t*.txt")

pub fn filter(&mut self) -> Option<String>

Gets the filter of the FileChooser

pub fn filter_value(&mut self) -> i32

Gets the current filename filter selection

pub fn set_filter_value(&mut self, f: i32)

Sets the filter for the dialog, can be: A single wildcard (eg. "*.txt"). Multiple wildcards (eg. "*.{cxx,h,H}"). A descriptive name followed by a \t and a wildcard (eg. "Text Files\t*.txt"). A list of separate wildcards with a \n between each (eg. "*.{cxx,H}\n*.txt"). A list of descriptive names and wildcards (eg. "C++ Files\t*.{cxx,H}\nTxt Files\t*.txt")

pub fn hide(&mut self)

Hides the file chooser

pub fn set_icon_size(&mut self, s: u8)

Sets the icon size of the FileChooser

pub fn icon_size(&mut self) -> u8

Gets the icon size of the FileChooser

pub fn set_label(&mut self, l: &str)

Sets the label of the FileChooser

pub fn label(&mut self) -> String

Gets the label of the FileChooser

pub fn set_ok_label(&mut self, l: &str)

Sets the label of the Ok button

pub fn ok_label(&mut self) -> String

Gets the label of the Ok button

pub fn set_preview(&mut self, e: bool)

Add preview to the FileChooser

pub fn preview(&self) -> bool

Returns whether preview is enabled for the FileChooser

pub fn rescan(&mut self)

Rescan the directory

pub fn rescan_keep_filename(&mut self)

Rescan the directory while keeping the file name

pub fn show(&mut self)

Shows the File Chooser

pub fn shown(&self) -> bool

Returns whether the file chooser is shown

pub fn set_text_color(&mut self, c: Color)

Sets the text color of the file chooser

pub fn text_color(&mut self) -> Color

Gets the text color of the file chooser

pub fn set_text_font(&mut self, f: Font)

Sets the text font of the file chooser

pub fn text_font(&mut self) -> Font

Gets the text font of the file chooser

pub fn set_text_size(&mut self, s: i32)

Sets the text size of the file chooser

pub fn text_size(&mut self) -> i32

Gets the text size of the file chooser

pub fn set_type(&mut self, t: FileChooserType)

Sets the type of the FileChooser

pub fn get_type(&mut self) -> FileChooserType

Gets the type of the FileChooser

pub unsafe fn user_data(&self) -> Option<Box<dyn FnMut()>>

Gets the user data of the FileChooser

Safety

Can invalidate the user data while the FileChooser is in use

pub fn value(&mut self, f: i32) -> Option<String>

Gets the file or dir name chosen by the FileChooser

pub fn set_value(&mut self, filename: &str)

Sets the file or dir name chosen by the FileChooser

pub fn visible(&mut self) -> bool

Returns whether the FileChooser is visible or not

pub fn window(&mut self) -> impl WindowExt

Return dialog window

pub fn set_add_favorites_label(msg: &'static str)

Set “Add favorites” label

pub fn set_all_files_label(msg: &'static str)

Set “All Files” label

pub fn set_custom_filter_label(msg: &'static str)

Set “Custom Filter” label

pub fn set_existing_file_label(msg: &'static str)

Set “Existing file” label

pub fn set_favorites_label(msg: &'static str)

Set “Favorites” label

pub fn set_filename_label(msg: &'static str)

Set “Filename” label

pub fn set_filesystems_label(msg: &'static str)

Set “Filesystems” label

pub fn set_manage_favorites_label(msg: &'static str)

Set “Manage favorites” label

pub fn set_new_directory_label(msg: &'static str)

Set “New directory” label

pub fn set_new_directory_tooltip(msg: &'static str)

Set “New directory” tooltip

pub fn set_preview_label(msg: &'static str)

Set “Preview” label

pub fn set_save_label(msg: &'static str)

Set “Save” label

pub fn set_show_label(msg: &'static str)

Set “Show” label

pub fn set_hidden_label(msg: &'static str)

Set “Hidden” label