fltk::enums

Enum FrameType

#[non_exhaustive]

#[repr(i32)]
pub enum FrameType {

    NoBox,
    FlatBox,
    UpBox,
    DownBox,
    UpFrame,
    DownFrame,
    ThinUpBox,
    ThinDownBox,
    ThinUpFrame,
    ThinDownFrame,
    EngravedBox,
    EmbossedBox,
    EngravedFrame,
    EmbossedFrame,
    BorderBox,
    ShadowBox,
    BorderFrame,
    ShadowFrame,
    RoundedBox,
    RShadowBox,
    RoundedFrame,
    RFlatBox,
    RoundUpBox,
    RoundDownBox,
    DiamondUpBox,
    DiamondDownBox,
    OvalBox,
    OShadowBox,
    OvalFrame,
    OFlatFrame,
    PlasticUpBox,
    PlasticDownBox,
    PlasticUpFrame,
    PlasticDownFrame,
    PlasticThinUpBox,
    PlasticThinDownBox,
    PlasticRoundUpBox,
    PlasticRoundDownBox,
    GtkUpBox,
    GtkDownBox,
    GtkUpFrame,
    GtkDownFrame,
    GtkThinUpBox,
    GtkThinDownBox,
    GtkThinUpFrame,
    GtkThinDownFrame,
    GtkRoundUpFrame,
    GtkRoundDownFrame,
    GleamUpBox,
    GleamDownBox,
    GleamUpFrame,
    GleamDownFrame,
    GleamThinUpBox,
    GleamThinDownBox,
    GleamRoundUpBox,
    GleamRoundDownBox,
    FreeBoxType,
    UserFrameType(UnmappedFrameType),

}

Defines the frame types which can be set using the set_frame() and set_down_frame() methods

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NoBox

No Box

FlatBox

Flat Box

UpBox

Up Box

DownBox

Down Box

UpFrame

Up Frame

DownFrame

Down Frame

ThinUpBox

Thin Up Box

ThinDownBox

Thin Down Box

ThinUpFrame

Thin Up Frame

ThinDownFrame

Thin Down Frame

EngravedBox

Engraved Box

EmbossedBox

Embossed Box

EngravedFrame

Engraved Frame

EmbossedFrame

Embossed Frame

BorderBox

Border Box

ShadowBox

Shadow Box

BorderFrame

Border Frame

ShadowFrame

Shadow Frame

RoundedBox

Rounded Box

RShadowBox

Rounded Shadow Box

RoundedFrame

Rounded Frame

RFlatBox

Rounded Flat Box

RoundUpBox

Round Up Box

RoundDownBox

Round Down Box

DiamondUpBox

Diamond Up Box

DiamondDownBox

Diamond Down Box

OvalBox

Oval Box

OShadowBox

Oval Shadow Box

OvalFrame

Oval Frame

OFlatFrame

Oval Flat Frame

PlasticUpBox

Plastic Up Box

PlasticDownBox

Plastic Down Box

PlasticUpFrame

Plastic Up Frame

PlasticDownFrame

Plastic Down Frame

PlasticThinUpBox

Plastic Thin Up Box

PlasticThinDownBox

Plastic Thin Down Box

PlasticRoundUpBox

Plastic Round Up Box

PlasticRoundDownBox

Plastic Round Down Box

GtkUpBox

Gtk Up Box

GtkDownBox

Gtk Down Box

GtkUpFrame

Gtk Up Frame

GtkDownFrame

Gtk Down Frame

GtkThinUpBox

Gtk Thin Up Box

GtkThinDownBox

Gtk Thin Down Box

GtkThinUpFrame

Gtk Thin Up Frame

GtkThinDownFrame

Gtk Thin Down Frame

GtkRoundUpFrame

Gtk Round Up Frame

GtkRoundDownFrame

Gtk Round Down Frame

GleamUpBox

Gleam Up Box

GleamDownBox

Gleam Down Box

GleamUpFrame

Gleam Up Frame

GleamDownFrame

Gleam Down Frame

GleamThinUpBox

Gleam Thin Up Box

GleamThinDownBox

Gleam Thin Down Box

GleamRoundUpBox

Gleam Round Up Box

GleamRoundDownBox

Gleam Round Down Box

FreeBoxType

Free BoxType

UserFrameType(UnmappedFrameType)

User-defined frame types

Implementations

impl FrameType

pub const OFlatBox: FrameType = FrameType::OFlatFrame

Alias OFlatFrame as OFlatBox

pub const GtkRoundDownBox: FrameType = FrameType::GtkRoundDownFrame

Alias GtkRoundDownFrame as GtkRoundDownBox

pub fn as_i32(&self) -> i32

Get the discriminant value or the user defined frame type

pub fn by_index(idx: usize) -> FrameType

Gets the Frame type by index

pub fn dx(self) -> i32

Get the frameโs x offset

pub fn dy(self) -> i32

Get the frameโs y offset

pub fn dw(self) -> i32

Get the frameโs width offset

pub fn dh(self) -> i32

Get the frameโs height offset

pub fn swap_frames(old_frame: FrameType, new_frame: FrameType)

Swap frames