fltk::enums

Struct Color

pub struct Color { /* private fields */ }

Defines colors used by FLTK. Colors are stored as RGBI values, the last being the index for FLTK colors in this enum. Colors in this enum don’t have an RGB stored. However, custom colors have an RGB, and don’t have an index. The RGBI can be acquired by casting the color to u32 and formatting it to 0x{08x}. The last 2 digits are the hexadecimal representation of the color in this enum. For example, Color::White, has a hex of 0x000000ff, ff being the 255 value of this enum. A custom color like Color::from_u32(0x646464), will have an representation as 0x64646400, of which the final 00 indicates that it is not stored in this enum. For convenience, the fmt::Display trait is implemented so that the name of the Color is shown when there is one, otherwise the RGB value is given.

Implementations

impl Color

pub const ForeGround: Color

ForeGround, label colors

pub const Foreground: Color

Foreground, label colors

pub const BackGround2: Color

BackGround2, Is the color inside input, output and text display widgets

pub const Background2: Color

Background2, Is the color inside input, output and text display widgets

pub const Inactive: Color

Inactive

pub const Selection: Color

Selection

pub const Free: Color

Free

pub const Gray0: Color

Gray0

pub const GrayRamp: Color

GrayRamp

pub const Dark3: Color

Dark3

pub const Dark2: Color

Dark2

pub const Dark1: Color

Dark1

pub const FrameDefault: Color

FrameDefault

pub const BackGround: Color

BackGround

pub const Background: Color

Background

pub const Light1: Color

Light1

pub const Light2: Color

Light2

pub const Light3: Color

Light3

pub const Black: Color

Black

pub const Red: Color

Red

pub const Green: Color

Green

pub const Yellow: Color

Yellow

pub const Blue: Color

Blue

pub const Magenta: Color

Magenta

pub const Cyan: Color

Cyan

pub const DarkRed: Color

DarkRed

pub const DarkGreen: Color

DarkGreen

pub const DarkYellow: Color

DarkYellow

pub const DarkBlue: Color

DarkBlue

pub const DarkMagenta: Color

DarkMagenta

pub const DarkCyan: Color

DarkCyan

pub const White: Color

White

pub const XtermBlack: Color

ANSI/xterm Black, not part of FLTK’s colormap

pub const XtermRed: Color

ANSI/xterm Red, not part of FLTK’s colormap

pub const XtermGreen: Color

ANSI/xterm Green, not part of FLTK’s colormap

pub const XtermYellow: Color

ANSI/xterm Yellow, not part of FLTK’s colormap

pub const XtermBlue: Color

ANSI/xterm Blue, not part of FLTK’s colormap

pub const XtermMagenta: Color

ANSI/xterm Magenta, not part of FLTK’s colormap

pub const XtermCyan: Color

ANSI/xterm Cyan, not part of FLTK’s colormap

pub const XtermWhite: Color

ANSI/xterm White, not part of FLTK’s colormap

pub const XtermBgRed: Color

ANSI/xterm background Red, not part of FLTK’s colormap

pub const XtermBgGreen: Color

ANSI/xterm background Green, not part of FLTK’s colormap

pub const XtermBgYellow: Color

ANSI/xterm background Yelllow, not part of FLTK’s colormap

pub const XtermBgBlue: Color

ANSI/xterm background Blue, not part of FLTK’s colormap

pub const XtermBgMagenta: Color

ANSI/xterm background Magenta, not part of FLTK’s colormap

pub const XtermBgCyan: Color

ANSI/xterm background Cyan, not part of FLTK’s colormap

pub const XtermBgWhite: Color

ANSI/xterm background White, not part of FLTK’s colormap

pub const TransparentBg: Color

Special background color value that lets the Terminal widget’s box() color show through behind the text. Not part of FLTK’s colormap

pub const fn bits(&self) -> u32

Gets the inner color representation

pub const fn from_rgb(r: u8, g: u8, b: u8) -> Color

Returns a color from RGB

pub const fn from_rgbi(rgbi: u32) -> Color

Returns a color enum from RGBI encoding

pub fn from_rgba_tuple(tup: (u8, u8, u8, u8)) -> Color

Create color from RGBA using alpha compositing. Works for non-group types.

pub const fn from_u32(val: u32) -> Color

Returns a color from hex or decimal

pub const fn from_hex(val: u32) -> Color

Returns a color from hex or decimal

pub fn from_hex_str(col: &str) -> Result<Color, FltkError>

Return a Color from a hex color format (#xxxxxx)

pub fn to_hex_str(&self) -> String

Returns the color in hex string format

pub fn by_index(idx: u8) -> Color

Returns a color by index of RGBI

pub fn inactive(&self) -> Color

Returns an inactive form of the color

pub fn darker(&self) -> Color

Returns an darker form of the color

pub fn lighter(&self) -> Color

Returns an lighter form of the color

pub fn gray_ramp(val: i32) -> Color

Returns a gray color value from black (i == 0) to white (i == FL_NUM_GRAY - 1)

pub fn color_average(c1: Color, c2: Color, weight: f32) -> Color

Returns a gray color value from black (i == 0) to white (i == FL_NUM_GRAY - 1)

pub fn contrast(fg: Color, bg: Color) -> Color

Returns a color that contrasts with the background color.

pub fn gray_scale(g: u8) -> Color

Returns the color closest to the passed grayscale value

pub fn rgb_color(r: u8, g: u8, b: u8) -> Color

Returns the color closest to the passed rgb value

pub fn to_rgb(&self) -> (u8, u8, u8)

Get the RGB value of the color