fltk::enums

Struct Font

pub struct Font { /* private fields */ }

Defines fonts used by FLTK

Implementations

impl Font

pub const Helvetica: Font

Helvetica

pub const HelveticaBold: Font

Helvetica Bold

pub const HelveticaItalic: Font

Helvetica Italic

pub const HelveticaBoldItalic: Font

Helvetica Bold Italic

pub const Courier: Font

Courier

pub const CourierBold: Font

Courier Bold

pub const CourierItalic: Font

Courier Italic

pub const CourierBoldItalic: Font

Courier Bold Italic

pub const Times: Font

Times

pub const TimesBold: Font

Times Bold

pub const TimesItalic: Font

Times Italic

pub const TimesBoldItalic: Font

Times Bold Italic

pub const Symbol: Font

Symbol

pub const Screen: Font

Screen

pub const ScreenBold: Font

Screen Bold

pub const Zapfdingbats: Font

Zapfdingbats

pub const fn bits(&self) -> i32

Gets the inner value of the Font

pub fn by_index(idx: usize) -> Font

Returns a font by index. This is the enum representation of the Font. If you change the default font for your app, which by default is Helvetica, Font::by_index(0) will still show Helvetica!

pub fn by_name(name: &str) -> Font

Gets the font by its name, can be queried via the app::get_font_names()

pub fn set_font(old: Font, new: &str)

Replace a current font with a loaded font

use fltk::enums::Font;
let font = Font::load_font("font.ttf").unwrap();
Font::set_font(Font::Helvetica, &font);

pub fn load_font<P: AsRef<Path>>(path: P) -> Result<String, FltkError>

Load font from file.

use fltk::enums::Font;
let font = Font::load_font("font.ttf").unwrap();
Font::set_font(Font::Helvetica, &font);

pub fn get_name(&self) -> String

Get the fontโs real name