fltk

Module group

Group widgets

Modules

experimental

Structs

ColorChooser

Creates a color chooser widget

Column

A wrapper around a Flex column

Flex

a Flexbox widget

Grid

Fltk’s grid widget

GridAlign

Defines alignment rules used by FLTK’s Grid

GridRange

Grid range

Group

Creates a group widget

HGrid

Defines a Horizontal Grid (custom widget). Requires setting the params manually using the set_params method, which takes the rows, columns and spacing.

Pack

Creates a widget pack

Row

A wrapper around a Flex row

Scroll

Creates a scroll group

Tabs

Creates a tab which can contain widgets

Tile

Creates a tile which can contain widgets. For the tiling to work correctly, the children of a Tile must cover the entire area of the widget, but not overlap. This means that all children must touch each other at their edges, and no gaps can be left inside the Tile. More info can be found here

VGrid

Defines a Vertical Grid (custom widget). Requires setting the params manually using the set_params method, which takes the rows, columns and spacing.

Wizard

Creates a wizard widget

Enums

FlexType

Defines Flex types

PackType

Defines pack types

ScrollType

Defines Scroll types

TabsOverflow

Defines how Tabs handle overflow