fltk::menu

Struct MacAppMenu

pub struct MacAppMenu;

Wrapper around Fl_Mac_App_Menu which exposes several static methods allowing the customization of the default system menu bar for the fltk application

Implementations

impl MacAppMenu

pub fn set_about(about: &'static str)

Sets the about text

pub fn set_print(print: &'static str)

Sets the print text

pub fn set_print_no_titlebar(print_no_titlebar: &'static str)

Sets the print no titlebar text

pub fn set_toggle_print_titlebar(toggle_print_titlebar: &'static str)

Sets the toggle print titlebar text

pub fn set_services(services: &'static str)

Sets the services text

pub fn set_hide(hide: &'static str)

Sets the hide text

pub fn set_hide_others(hide_others: &'static str)

Sets the hide others text

pub fn set_show(show: &'static str)

Sets the show text

pub fn set_quit(quit: &'static str)

Sets the quit text

pub fn custom_application_menu_items(m: MenuItem)

Adds custom menu items to the application menu of the system menu bar. They are positioned after the โPrint Front Window / Toggle printing of titlebarโ items, or at their place if an item is removed by providing empty text