fltk::menu

Struct MenuItem

pub struct MenuItem { /* private fields */ }

Creates a menu item

Implementations

impl MenuItem

pub unsafe fn from_ptr(ptr: *mut Fl_Menu_Item) -> MenuItem

Initializes a MenuItem from a pointer

Safety

The pointer must be valid

pub unsafe fn as_ptr(&self) -> *mut Fl_Menu_Item

Returns the inner pointer from a MenuItem

Safety

Can return multiple mutable pointers to the same item

pub fn new(choices: &[&'static str]) -> MenuItem

Initializes a new menu item. This will allocate a static MenuItem, that is expected to live for the entirety of the program.

pub fn popup(&self, x: i32, y: i32) -> Option<MenuItem>

Creates a popup menu at the specified coordinates and returns its choice

pub fn pulldown( &self, x: i32, y: i32, w: i32, h: i32, picked: Option<MenuItem>, menu: Option<&impl MenuExt>, ) -> Option<MenuItem>

Creates a pulldown menu at the specified coordinates and returns its choice

pub fn label(&self) -> Option<String>

Returns the label of the menu item

pub fn set_label(&mut self, txt: &str)

Sets the label of the menu item

pub fn label_type(&self) -> LabelType

Returns the label type of the menu item

pub fn set_label_type(&mut self, typ: LabelType)

Sets the label type of the menu item

pub fn label_color(&self) -> Color

Returns the label color of the menu item

pub fn set_label_color(&mut self, color: Color)

Sets the label color of the menu item

pub fn label_font(&self) -> Font

Returns the label font of the menu item

pub fn set_label_font(&mut self, font: Font)

Sets the label font of the menu item

pub fn label_size(&self) -> i32

Returns the label size of the menu item

pub fn set_label_size(&mut self, sz: i32)

Sets the label size of the menu item

pub fn value(&self) -> bool

Returns the value of the menu item

pub fn set(&mut self)

Sets the menu item

pub fn clear(&mut self)

Turns the check or radio item “off” for the menu item

pub fn visible(&self) -> bool

Returns whether the menu item is visible or not

pub fn active(&self) -> bool

Returns whether the menu item is active

pub fn activate(&mut self)

Activates the menu item

pub fn deactivate(&mut self)

Deactivates the menu item

pub fn is_submenu(&self) -> bool

Returns whether a menu item is a submenu

pub fn is_checkbox(&self) -> bool

Returns whether a menu item is a checkbox

pub fn is_radio(&self) -> bool

Returns whether a menu item is a radio item

pub fn show(&mut self)

Shows the menu item

pub fn hide(&mut self)

Hides the menu item

pub fn next(&self, idx: i32) -> Option<MenuItem>

Get the next menu item skipping submenus

pub fn children(&self) -> i32

Get children of MenuItem

pub fn submenus(&self) -> i32

Get the submenu count

pub fn size(&self) -> i32

Get the size of the MenuItem

pub fn at(&self, idx: i32) -> Option<MenuItem>

Get the menu item at idx

pub unsafe fn user_data(&self) -> Option<Box<dyn FnMut()>>

Get the user data

Safety

Can return multiple mutable instances of the user data, which has a different lifetime than the object

pub fn set_callback<F: FnMut(&mut Choice) + 'static>(&mut self, cb: F)

Set a callback for the menu item

pub fn do_callback<W: MenuExt>(&mut self, w: &W)

Run the menu item’s callback

pub fn emit<T: 'static + Clone + Send + Sync>( &mut self, sender: Sender<T>, msg: T, )

Use a sender to send a message during callback

pub fn was_deleted(&self) -> bool

Check if a menu item was deleted

pub fn draw<M: MenuExt>( &self, x: i32, y: i32, w: i32, h: i32, menu: &M, selected: bool, )

Draw a box around the menu item. Requires the call to be made inside a MenuExt-implementing widget’s own draw method

pub fn measure(&self) -> (i32, i32)

Measure the width and height of a menu item

pub fn add_image<I: ImageExt>(&mut self, image: Option<I>, on_left: bool)

Add an image to a menu item

use fltk::{prelude::*, *};
const PXM: &[&str] = &[
    "13 11 3 1",
    "   c None",
    "x  c #d8d833",
    "@  c #808011",
    "             ",
    "     @@@@    ",
    "    @xxxx@   ",
    "@@@@@xxxx@@  ",
    "@xxxxxxxxx@  ",
    "@xxxxxxxxx@  ",
    "@xxxxxxxxx@  ",
    "@xxxxxxxxx@  ",
    "@xxxxxxxxx@  ",
    "@xxxxxxxxx@  ",
    "@@@@@@@@@@@  "
];
let image = image::Pixmap::new(PXM).unwrap();
let mut menu = menu::MenuBar::default();
menu.add(
    "&File/Open...\t",
    enums::Shortcut::Ctrl | 'o',
    menu::MenuFlag::Normal,
    |_| println!("Opened file!"),
);
if let Some(mut item) = menu.find_item("&File/Open...\t") {
    item.add_image(Some(image), true);
}

pub fn add<F: FnMut(&mut Choice) + 'static>( &mut self, name: &str, shortcut: Shortcut, flag: MenuFlag, cb: F, ) -> i32

Add a menu item

pub fn insert<F: FnMut(&mut Choice) + 'static>( &mut self, idx: i32, name: &str, shortcut: Shortcut, flag: MenuFlag, cb: F, ) -> i32

Insert a menu item

pub fn add_emit<T: 'static + Clone + Send + Sync>( &mut self, label: &str, shortcut: Shortcut, flag: MenuFlag, sender: Sender<T>, msg: T, ) -> i32

Add a menu item along with an emit (sender and message).

pub fn insert_emit<T: 'static + Clone + Send + Sync>( &mut self, idx: i32, label: &str, shortcut: Shortcut, flag: MenuFlag, sender: Sender<T>, msg: T, ) -> i32

Insert a menu item along with an emit (sender and message).

pub fn set_shortcut(&mut self, shortcut: Shortcut)

Set the menu item’s shortcut

pub fn set_flag(&mut self, flag: MenuFlag)

Set the menu item’s shortcut