fltk::misc

Struct Tooltip

pub struct Tooltip {}

Controls tooltips on an application-wide basis; use .set_tooltip() to add a tooltip to a particular widget

Implementations

impl Tooltip

pub fn delay() -> f32

Gets the tooltips delay

pub fn set_delay(f: f32)

Sets the tooltips delay

pub fn hidedelay() -> f32

Gets the tooltips hide delay

pub fn set_hidedelay(f: f32)

Sets the tooltips hide delay

pub fn hoverdelay() -> f32

Gets the tooltips hover delay

pub fn set_hoverdelay(f: f32)

Sets the tooltips hover delay

pub fn enabled() -> bool

Returns whether the tooltips are enabled

pub fn enable(b: bool)

Sets tooltips to be displayed if b is true; otherwise not to be displayed

pub fn disable()

Disables the display of all tooltips

pub fn enter_area<W: WidgetExt>( widget: &W, x: i32, y: i32, w: i32, h: i32, tip: &'static CStr, )

Used to provide tooltips for internal pieces of your widget. Check FLTKโs documentation. The text of the tooltip must be a static CStr since the data is not copied by FLTK. This also avoid memory leaks in user code.

pub fn current_widget() -> impl WidgetExt

Returns the current widget associated with the tooltip

pub fn current<W: WidgetExt>(w: &W)

Sets the current widget associated with the tooltip

pub fn font() -> Font

Gets the tooltips font

pub fn set_font(font: Font)

Sets the tooltips font

pub fn font_size() -> i32

Gets the tooltips font size

pub fn set_font_size(s: i32)

Sets the tooltips font size

pub fn color() -> Color

Gets the tooltips color

pub fn set_color(c: Color)

Sets the tooltips color

pub fn text_color() -> Color

Gets the tooltips text color

pub fn set_text_color(c: Color)

Sets the tooltips text color

pub fn margin_width() -> i32

Gets the tooltips margin width

pub fn set_margin_width(v: i32)

Sets the tooltips margin width

pub fn margin_height() -> i32

Gets the tooltips margin height

pub fn set_margin_height(v: i32)

Sets the tooltips margin height

pub fn wrap_width() -> i32

Gets the tooltips wrap width

pub fn set_wrap_width(v: i32)

Sets the tooltips wrap width

pub fn current_window() -> impl WindowExt

Returns the window used for tooltips