fltk

Module prelude

All fltk widget traits and flt error types

Re-exports

pub use crate::app::WidgetId;

Enums

FltkError

Error types returned by fltk-rs + wrappers of std errors

FltkErrorKind

Error kinds enum for FltkError

Traits

BrowserExt

Defines the methods implemented by all browser types More info can be found in the wiki

ButtonExt

Defines the methods implemented by all button widgets. More details can be found in the wiki.

DisplayExt

Defines the methods implemented by TextDisplay and TextEditor More details can be found in the wiki.

GroupExt

Defines the methods implemented by all group widgets. These widgets include Window types and others found in the group module: Group, Scroll, Pack, Tile, Flex …etc. Widgets implementing the GroupExt trait, are characterized by having to call ::end() method to basically close them. More details can be found in the wiki.

ImageExt

Defines the methods implemented by all image types

InputExt

Defines the methods implemented by all input and output widgets. More details can be found in the wiki.

MenuExt

Defines the methods implemented by all menu widgets These are found in the menu module: MenuBar, SysMenuBar, Choice, MenuButton …etc. Menus function in 2 main ways which are discussed in the wiki

SurfaceDevice

Defines the methods implemented by all surface types, currently ImageSurface

TableExt

Defines the methods implemented by table types. More details can be found in the wiki.

ValuatorExt

Defines the methods implemented by all valuator widgets More details can be found in the wiki.

WidgetBase

Defines the extended methods implemented by all widgets

WidgetExt

Defines the methods implemented by all widgets

WidgetType

A trait defined for all enums passable to the WidgetExt::set_type() method

WindowExt

Defines the methods implemented by all window widgets. More details can be found in the wiki. Windows (which can be found in the window module) implement GroupExt as well.