Trait fltk::prelude::ImageExt

pub unsafe trait ImageExt {

    fn copy(&self) -> Self
    where
        Self: Sized;

    fn draw(&mut self, x: i32, y: i32, width: i32, height: i32);

    fn draw_ext(
        &mut self, 
        x: i32, 
        y: i32, 
        width: i32, 
        height: i32, 
        cx: i32, 
        cy: i32
    );

    fn width(&self) -> i32;

    fn height(&self) -> i32;

    fn w(&self) -> i32;

    fn h(&self) -> i32;

    unsafe fn as_image_ptr(&self) -> *mut Fl_Image;

    unsafe fn from_image_ptr(ptr: *mut Fl_Image) -> Self
    where
        Self: Sized;

    fn to_rgb_data(&self) -> Vec<u8>;

    fn to_raw_data(&self) -> *const *const u8;

    fn to_rgb(&self) -> Result<RgbImage, FltkError>;

    fn scale(
        &mut self, 
        width: i32, 
        height: i32, 
        proportional: bool, 
        can_expand: bool
    );

    fn count(&self) -> i32;

    fn data_w(&self) -> i32;

    fn data_h(&self) -> i32;

    fn depth(&self) -> ColorDepth;

    fn ld(&self) -> i32;

    fn inactive(&mut self);

    unsafe fn delete(img: Self)
    where
        Self: Sized;

    fn was_deleted(&self) -> bool;

    unsafe fn increment_arc(&mut self);

    unsafe fn decrement_arc(&mut self);

    unsafe fn into_image<I: ImageExt>(self) -> I
    where
        Self: Sized;

}

Defines the methods implemented by all image types

Required methods

fn copy(&self) -> Self where
Self: Sized,

Performs a deep copy of the image

fn draw(&mut self, x: i32, y: i32, width: i32, height: i32)

Draws the image at the presupplied coordinates and size

fn draw_ext(
&mut self,
x: i32,
y: i32,
width: i32,
height: i32,
cx: i32,
cy: i32
)

Draws the image at the presupplied coordinates and size and offset cx, cy

fn width(&self) -> i32

Return the width of the image

fn height(&self) -> i32

Return the height of the image

fn w(&self) -> i32

Return the width of the image

fn h(&self) -> i32

Return the height of the image

unsafe fn as_image_ptr(&self) -> *mut Fl_Image

Retunrs a pointer of the image

Safety

Can return multiple mutable pointers to the image

unsafe fn from_image_ptr(ptr: *mut Fl_Image) -> Self where
Self: Sized,

Transforms a raw image pointer to an image

Safety

The pointer must be valid

fn to_rgb_data(&self) -> Vec<u8>

Returns the underlying raw rgb image data

fn to_raw_data(&self) -> *const *const u8

Returns the underlying raw image data

fn to_rgb(&self) -> Result<RgbImage, FltkError>

Transforms the image into an RgbImage

Errors

Errors on failure to transform to RgbImage

fn scale(
&mut self,
width: i32,
height: i32,
proportional: bool,
can_expand: bool
)

Scales the image

fn count(&self) -> i32

Return the count of pointers in an image (Pixmaps have more than 1, bitmaps have 0, Rgb based images have 1)

fn data_w(&self) -> i32

Gets the image’s data width

fn data_h(&self) -> i32

Gets the image’s data height

fn depth(&self) -> ColorDepth

Gets the image’s depth

fn ld(&self) -> i32

Gets the image’s line data size

fn inactive(&mut self)

Greys the image

unsafe fn delete(img: Self) where
Self: Sized,

Deletes the image

Safety

An image shouldn’t be deleted while it’s being used by a widget

fn was_deleted(&self) -> bool

Checks if the image was deleted

unsafe fn increment_arc(&mut self)

INTERNAL: Manually increment the atomic refcount

Safety

The underlying image pointer must be valid

unsafe fn decrement_arc(&mut self)

INTERNAL: Manually decrement the atomic refcount

Safety

The underlying image pointer must be valid

unsafe fn into_image<I: ImageExt>(self) -> I where
Self: Sized,

Transforms an Image base into another Image

Safety

Can be unsafe if used to downcast to an image of different format