Trait fltk::prelude::InputExt

pub unsafe trait InputExt: WidgetExt {

    fn value(&self) -> String;

    fn set_value(&mut self, val: &str);

    fn maximum_size(&self) -> i32;

    fn set_maximum_size(&mut self, val: i32);

    fn position(&self) -> i32;

    fn set_position(&mut self, val: i32) -> Result<(), FltkError>;

    fn mark(&self) -> i32;

    fn set_mark(&mut self, val: i32) -> Result<(), FltkError>;

    fn replace(
        &mut self, 
        beg: i32, 
        end: i32, 
        val: &str
    ) -> Result<(), FltkError>;

    fn insert(&mut self, txt: &str) -> Result<(), FltkError>;

    fn append(&mut self, txt: &str) -> Result<(), FltkError>;

    fn copy(&mut self) -> Result<(), FltkError>;

    fn undo(&mut self) -> Result<(), FltkError>;

    fn cut(&mut self) -> Result<(), FltkError>;

    fn text_font(&self) -> Font;

    fn set_text_font(&mut self, font: Font);

    fn text_color(&self) -> Color;

    fn set_text_color(&mut self, color: Color);

    fn text_size(&self) -> i32;

    fn set_text_size(&mut self, sz: i32);

    fn readonly(&self) -> bool;

    fn set_readonly(&mut self, val: bool);

    fn wrap(&self) -> bool;

    fn set_wrap(&mut self, val: bool);

}

Defines the methods implemented by all input and output widgets

Required methods

fn value(&self) -> String

Returns the value inside the input/output widget

fn set_value(&mut self, val: &str)

Sets the value inside an input/output widget

fn maximum_size(&self) -> i32

Returns the maximum size (in bytes) accepted by an input/output widget

fn set_maximum_size(&mut self, val: i32)

Sets the maximum size (in bytes) accepted by an input/output widget

fn position(&self) -> i32

Returns the index position inside an input/output widget

fn set_position(&mut self, val: i32) -> Result<(), FltkError>

Sets the index postion inside an input/output widget

Errors

Errors on failure to set the cursor position in the text

fn mark(&self) -> i32

Returns the index mark inside an input/output widget

fn set_mark(&mut self, val: i32) -> Result<(), FltkError>

Sets the index mark inside an input/output widget

Errors

Errors on failure to set the mark

fn replace(&mut self, beg: i32, end: i32, val: &str) -> Result<(), FltkError>

Replace content with a &str

Errors

Errors on failure to replace text

fn insert(&mut self, txt: &str) -> Result<(), FltkError>

Insert a &str

Errors

Errors on failure to insert text

fn append(&mut self, txt: &str) -> Result<(), FltkError>

Append a &str

Errors

Errors on failure to append text

fn copy(&mut self) -> Result<(), FltkError>

Copy the value within the widget

Errors

Errors on failure to copy selection

fn undo(&mut self) -> Result<(), FltkError>

Undo changes

Errors

Errors on failure to undo

fn cut(&mut self) -> Result<(), FltkError>

Cut the value within the widget

Errors

Errors on failure to cut selection

fn text_font(&self) -> Font

Return the text font

fn set_text_font(&mut self, font: Font)

Sets the text font

fn text_color(&self) -> Color

Return the text color

fn set_text_color(&mut self, color: Color)

Sets the text color

fn text_size(&self) -> i32

Return the text size

fn set_text_size(&mut self, sz: i32)

Sets the text size

fn readonly(&self) -> bool

Returns whether the input/output widget is readonly

fn set_readonly(&mut self, val: bool)

Set readonly status of the input/output widget

fn wrap(&self) -> bool

Return whether text is wrapped inside an input/output widget

fn set_wrap(&mut self, val: bool)

Set whether text is wrapped inside an input/output widget