fltk::surface

Struct SvgFileSurface

pub struct SvgFileSurface { /* private fields */ }

An SVG image surface object which allows drawing to an svg file Example usage:

use fltk::{prelude::*, *};
let but = button::Button::new(0, 0, 80, 40, "Click");
// We need the destructor of SvgFileSurface to actually create the image
{
    let sur = surface::SvgFileSurface::new(but.width(), but.height(), "temp.svg");
    surface::SvgFileSurface::push_current(&sur);
    draw::set_draw_color(enums::Color::White);
    draw::draw_rectf(0, 0, but.width(), but.height());
    sur.draw(&but, 0, 0);
    surface::SvgFileSurface::pop_current();
}

Implementations

impl SvgFileSurface

pub fn new<P: AsRef<Path>>(width: i32, height: i32, path: P) -> SvgFileSurface

Returns a new SvgFileSurface. The path if non-existent will be created

Panics

Panics on File creation failure and on writablity failure

pub fn set_origin(&mut self, x: i32, y: i32)

Sets the origin of the SvgFileSurface

pub fn printable_rect(&self) -> (i32, i32)

Returns the width and height of the printable rect

pub fn draw<W: WidgetExt>(&self, widget: &W, delta_x: i32, delta_y: i32)

Draw a widget in an svg file surface. The .svg file is not complete until the destructor was run

pub fn draw_decorated_window<W: WindowExt>( &self, win: &W, x_offset: i32, y_offset: i32, )

draw a decorated window