fltk::terminal

Struct Utf8Char

pub struct Utf8Char { /* private fields */ }

Class to manage the terminal’s individual UTF-8 characters. Includes fg/bg color, attributes (BOLD, UNDERLINE..) This is a low-level “protected” class in the fltk library

Implementations

impl Utf8Char

pub fn new(c: u8) -> Self

Construct a new Utf8Char, single-byte only. This is really only useful for testing. ‘c’ must be “printable” ASCII in the range (0x20 <= c <= 0x7e). Anything outside of that is silently ignored.

Allocated Utf8Char will never be deleted.

pub fn attr_bgcolor(&self, term: Option<&Terminal>) -> Color

Return the actual displayed color of char u8c possibly influenced by BOLD or DIM if the char is from Xterm. BG color will be derived from the widget color if a widget is specified and the color is TransparentBg, and that won’t be influenced by charflag attributes.

pub fn attr_fgcolor(&self, term: Option<&Terminal>) -> Color

Return the actual displayed fg color of char u8c possibly influenced by BOLD or DIM if the char is from Xterm. If a term widget is specified (i.e. not None), don’t let the color be influenced by the attribute bits if it matches the term widget’s own color().

pub fn attrib(&self) -> Attrib

Return the attributes for this character.

pub fn bgcolor(&self) -> Color

Return the background color for this character.

pub fn fgcolor(&self) -> Color

Return the foreground color for this character.

pub fn charflags(&self) -> CharFlags

Return the xterm CharFlags bits

pub fn is_char(&self, c: u8) -> bool

Returns true if the character text in this struct matches the given ASCII character

pub fn length(&self) -> usize

Return the length of this character in bytes (UTF-8 can be multibyte)

pub fn max_utf8(&self) -> usize

Return the maximum length in bytes of a UTF-8 character

pub fn pwidth(&self) -> f64

Return the width of this character in floating point pixels.

WARNING: Uses current font, so assumes font and font_size have already been set to current font!

pub fn pwidth_int(&self) -> usize

Return the width of this character in integer pixels.

WARNING: Uses current font, so assumes font and font_size have already been set to current font!

pub fn text_utf8(&self) -> &[u8]

Return the UTF-8 text string for this character.