fltk

Module text

Text display widgets

Structs

SimpleTerminalDeprecated

Creates an editable text display widget to handle terminal-like behavior, such as logging events or debug information. SimpleTerminal already has an internal buffer. It is NOT is a full terminal emulator; it does NOT handle stdio redirection, pipes, pseudo ttys, termio character cooking, keyboard input processing, screen addressing, random cursor positioning, curses compatibility, or VT100/xterm emulation.

StyleTableEntry

Defines the styles used in the set_highlight_data, which is used with style buffers

StyleTableEntryExt

Defines the styles used in the set_highlight_data, which is used with style buffers

TextBuffer

Wraps a text buffer, Cloning a text buffer invalidates the underlying pointer, thus the no derive(Clone)

TextDisplay

Creates a non-editable text display widget

TextEditor

Creates an editable text display widget

Enums

Cursor

Defines the text cursor styles supported by fltk

DragType

Defines drag types

TextAttr

The attribute of the style entry

WrapMode

Defines wrap modes

Type Aliases

TextEditorPtr

Alias Fl_Text_Editor for use in add_key_binding