Struct fltk::text::TextEditor

pub struct TextEditor { /* fields omitted */ }

Creates an editable text display widget

Implementations

impl TextEditor

pub fn set_insert_mode(&mut self, b: bool)

Set to insert mode

pub fn insert_mode(&self) -> bool

Returns whether insert mode is set

pub fn set_tab_nav(&mut self, val: bool)

Set tab navigation

pub fn tab_nav(&self) -> bool

Returns whether tab navigation is set

pub fn copy(&self)

Copies the text within the TextEditor widget

pub fn cut(&self)

Cuts the text within the TextEditor widget

pub fn paste(&self)

Pastes text from the clipboard into the TextEditor widget

pub fn undo(&self)

Undo changes in the TextEditor widget

pub fn kf_default(&mut self, c: Key)

Inserts the text associated with key ‘c’

pub fn kf_ignore(&mut self, c: Key)

Ignores the key ‘c’ in editor

pub fn kf_backspace(&mut self)

Does a backspace

pub fn kf_enter(&mut self)

Inserts a new line

pub fn kf_move(&mut self, c: Key)

Moves the cursor in the direction indicated by the key

pub fn kf_shift_move(&mut self, c: Key)

Extends the current selection in the direction of key ‘c’

pub fn kf_ctrl_move(&mut self, c: Key)

Moves the current text cursor in the direction indicated by control key ‘c’

pub fn kf_c_s_move(&mut self, c: Key)

Extends the current selection in the direction indicated by control key ‘c’

pub fn kf_meta_move(&mut self, c: Key)

Moves the current text cursor in the direction indicated by meta key ‘c’

pub fn kf_m_s_move(&mut self, c: Key)

Extends the current selection in the direction indicated by meta key ‘c’

pub fn kf_home(&mut self)

Moves the text cursor to the beginning of the current line

pub fn kf_end(&mut self)

Moves the text cursor to the end of the current line

pub fn kf_left(&mut self)

Moves the text cursor one character to the left

pub fn kf_up(&mut self)

Moves the text cursor one line up

pub fn kf_right(&mut self)

Moves the text cursor one character to the right

pub fn kf_down(&mut self)

Moves the text cursor one line down

pub fn kf_page_up(&mut self)

Moves the text cursor up one page

pub fn kf_page_down(&mut self)

Moves the text cursor down one page

pub fn kf_insert(&mut self)

Toggles the insert mode for the editor

pub fn kf_delete(&mut self)

Does a delete of selected text or the current character in the current buffer

pub fn kf_select_all(&mut self)

Selects all text in the associated buffer