fltk

Module window

Window widgets Window widgets

Multithreaded applications should always create/show/open/close windows from the main thread (This might or might not work on your specific target, due to fltk calling the underlying platform’s window code. If you want portability, avoid it.) If you need to trigger showing a windows from another thread, use messages to notify the main thread that the window needs showing. An alternative to that is awake_callback

Structs

AndroidWindow

An Android window

DoubleWindow

Creates a double (buffered) window widget

MenuWindow

Creates a Menu window widget

OverlayWindow

Creates an overlay (buffered) window widget

SingleWindow

Creates a single (buffered) window widget

Enums

WindowType

Defines the window type

Type Aliases

GlContext

A wrapper around a raw OpenGL context

RawHandle

Opaque raw window handle (*mut c_void to HWND on Windows and NSWindow on macOS), XID (u64) raw window handle for X11

Window

Creates a window widget