fltk::window

Struct AndroidWindow

pub struct AndroidWindow { /* private fields */ }

An Android window

Implementations

impl AndroidWindow

pub fn with_pos(self, x: i32, y: i32) -> Self

Initialize to position x, y

pub fn with_size(self, width: i32, height: i32) -> Self

Initialize to size width, height

pub fn with_label(self, title: &str) -> Self

Initialize with a label

pub fn with_align(self, align: Align) -> Self

Initialize with alignment

pub fn with_type<T: WidgetType>(self, typ: T) -> Self

Initialize with type

pub fn below_of<W: WidgetExt>(self, wid: &W, padding: i32) -> Self

Initialize at bottom of another widget

pub fn above_of<W: WidgetExt>(self, wid: &W, padding: i32) -> Self

Initialize above of another widget

pub fn right_of<W: WidgetExt>(self, wid: &W, padding: i32) -> Self

Initialize right of another widget

pub fn left_of<W: WidgetExt>(self, wid: &W, padding: i32) -> Self

Initialize left of another widget

pub fn center_of<W: WidgetExt>(self, w: &W) -> Self

Initialize center of another widget

pub fn center_x<W: WidgetExt>(self, w: &W) -> Self

Initialize center of another widget on the x axis

pub fn center_y<W: WidgetExt>(self, w: &W) -> Self

Initialize center of another widget on the y axis

pub fn center_of_parent(self) -> Self

Initialize center of parent

pub fn size_of<W: WidgetExt>(self, w: &W) -> Self

Initialize to the size of another widget

pub fn size_of_parent(self) -> Self

Initialize to the size of the parent