fltk

Module dialog

Dialog widgets

Structs

FileChooser

FLTK’s own FileChooser. Which differs for the Native FileDialog Example:

FileChooserType

The types of FileChooser

FileDialog

FLTK’s NativeFileChooser

FileDialogOptions

Defines the File dialog options, which can be set using the set_option() method.

HelpDialog

Creates a help dialog

Enums

BeepType

Defines the type of beep to be passed to the beep function

ColorMode

Color modes to be used with the color chooser

FileDialogAction

Rusult of try_show

FileDialogType

Defines the type of dialog, which can be changed dynamically using the set_type() method

Functions

alert

Displays an alert box

alert_default

Displays an alert box, the dialog is positioned at the pointer hotspot

beep

Emits a beep

choiceDeprecated

Displays a choice box with up to three choices. Choosing a value returns its index from the arguments

choice2

Displays a choice box with up to three choices. Closing the dialog returns None. Choosing a value returns its index from the arguments.

choice2_default

Displays a choice box with up to three choices. An empty choice will not be shown. Closing the dialog returns None. Choosing a value returns its index from the arguments. The dialog is positioned at the pointer hotspot

choice_defaultDeprecated

Displays a choice box with up to three choices. The dialog is positioned at the pointer hotspot

color_chooser

Spawns a color chooser dialog.

color_chooser_with_default

Spawns a color chooser dialog.

dir_chooser

Shows a directory chooser returning a String

file_chooser

Shows a file chooser returning a String. The pattern field takes the same argument the FileChooser::set_filter method. Example:

input

Displays an input box, which returns the inputted string. Can be used for gui io

input_default

Displays an input box, which returns the inputted string. Can be used for gui io. The dialog is positioned at the pointer hotspot

message

Displays a message box

message_default

Displays a message box, the dialog is positioned at the pointer hotspot

message_hotspot

Get whether hotspot is enabled for FLTK’s dialog boxes

message_icon

Get the frame holding the icon of FLTK’s dialog boxes

message_icon_label

Set the next dialog’s icon label

message_set_font

Set the font and font size of FLTK’s dialog boxes

message_set_hotspot

Set whether hotspot is enabled for FLTK’s dialog boxes

message_title

Set the next dialog’s title

message_title_default

Set the default title for a dialog

password

Shows an input box, but with hidden string

password_default

Shows an input box, but with hidden string. The dialog is positioned at the pointer hotspot

Type Aliases

NativeFileChooser

FLTK’s NativeFileChooser

NativeFileChooserAction

Alias to FileDialogAction

NativeFileChooserOptions

Alias to FileDialogOptions

NativeFileChooserType

Alias for NativeFileChooserType