fltk::prelude

Trait WidgetBase

pub unsafe trait WidgetBase: WidgetExt {
    // Required methods
    fn new<'a, T: Into<Option<&'a str>>>(
        x: i32,
        y: i32,
        width: i32,
        height: i32,
        title: T,
    ) -> Self;
    fn default_fill() -> Self;
    fn delete(wid: Self)
       where Self: Sized;
    unsafe fn from_widget_ptr(ptr: *mut Fl_Widget) -> Self;
    unsafe fn from_widget<W: WidgetExt>(w: W) -> Self;
    fn handle<F: FnMut(&mut Self, Event) -> bool + 'static>(&mut self, cb: F);
    fn draw<F: FnMut(&mut Self) + 'static>(&mut self, cb: F);
    fn resize_callback<F: FnMut(&mut Self, i32, i32, i32, i32) + 'static>(
        &mut self,
        cb: F,
    );

    // Provided methods
    unsafe fn assume_derived(&mut self) { ... }
    fn from_dyn_widget<W: WidgetExt>(_w: &W) -> Option<Self>
       where Self: Sized { ... }
    fn from_dyn_widget_ptr(_w: *mut Fl_Widget) -> Option<Self>
       where Self: Sized { ... }
}

Defines the extended methods implemented by all widgets

Safety

fltk-rs traits depend on some FLTK internal code

Warning

fltk-rs traits are non-exhaustive, to avoid future breakage if you try to implement them manually, use the Deref and DerefMut pattern or the widget_extends! macro

Required Methods

fn new<'a, T: Into<Option<&'a str>>>( x: i32, y: i32, width: i32, height: i32, title: T, ) -> Self

Creates a new widget, takes an x, y coordinates, as well as a width and height, plus a title

Arguments

x - The x coordinate in the screen

y - The y coordinate in the screen

width - The width of the widget

heigth - The height of the widget

title - The title or label of the widget

To use dynamic strings use with_label(self, &str) or set_label(&mut self, &str). labels support special symbols preceded by an @ sign and for the associated formatting.

fn default_fill() -> Self

Constructs a widget with the size of its parent

fn delete(wid: Self)
where Self: Sized,

Deletes widgets and their children.

unsafe fn from_widget_ptr(ptr: *mut Fl_Widget) -> Self

transforms a widget pointer to a Widget, for internal use

Safety

The pointer must be valid

unsafe fn from_widget<W: WidgetExt>(w: W) -> Self

Get a widget from base widget

Safety

The underlying object must be valid

fn handle<F: FnMut(&mut Self, Event) -> bool + 'static>(&mut self, cb: F)

Set a custom handler, where events are managed manually, akin to Fl_Widget::handle(int). Handled or ignored events should return true, unhandled events should return false. takes the widget as a closure argument. The ability to handle an event might depend on handling other events, as explained here

fn draw<F: FnMut(&mut Self) + 'static>(&mut self, cb: F)

Set a custom draw method. takes the widget as a closure argument. macOS requires that WidgetBase::draw actually calls drawing functions

fn resize_callback<F: FnMut(&mut Self, i32, i32, i32, i32) + 'static>( &mut self, cb: F, )

Perform a callback on resize. Avoid resizing the parent or the same widget to avoid infinite recursion

Provided Methods

unsafe fn assume_derived(&mut self)

Makes the widget derived

Safety

Calling this on a non-derived widget can cause undefined behavior

fn from_dyn_widget<W: WidgetExt>(_w: &W) -> Option<Self>
where Self: Sized,

Cast a type-erased widget back to its original widget

fn from_dyn_widget_ptr(_w: *mut Fl_Widget) -> Option<Self>
where Self: Sized,

Cast a type-erased widget pointer back to its original widget

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.