fltk::printer

Struct Printer

pub struct Printer { /* private fields */ }

Defines a printer object. Example usage:

use fltk::{prelude::*, *};
let mut but = button::Button::default();
but.set_callback(|widget| {
    let mut printer = printer::Printer::default();
    if printer.begin_job(1).is_ok() {
        printer.begin_page().ok();
        let (width, height) = printer.printable_rect();
        draw::set_draw_color(enums::Color::Black);
        draw::set_line_style(draw::LineStyle::Solid, 2);
        draw::draw_rect(0, 0, width, height);
        draw::set_font(enums::Font::Courier, 12);
        printer.set_origin(width / 2, height / 2);
        printer.print_widget(widget, -widget.width() / 2, -widget.height() / 2);
        printer.end_page().ok();
        printer.end_job();
    }
});

Implementations

impl Printer

pub fn begin_job( &mut self, pagecount: i32, ) -> Result<(Option<i32>, Option<i32>), FltkError>

Begins a print job. pagecount The total number of pages to be created. Use 0 if this number is unknown Returns a tuple (frompage, topage) indicating the chosen pages by the user

Errors

Errors on failure to print

pub fn end_page(&mut self) -> Result<(), FltkError>

End the print page

Errors

Errors on failure to end the page

pub fn end_job(&mut self)

Ends the print job

pub fn begin_page(&mut self) -> Result<(), FltkError>

Begins a print page

Errors

Errors on failure to begin the page

pub fn printable_rect(&self) -> (i32, i32)

Returns the width and height of the printable rect

pub fn margins(&self) -> (i32, i32, i32, i32)

Returns the coordinates of the printable margins. (left, top, right, bottom)

pub fn origin(&self) -> (i32, i32)

Get the origin coordinates of the printable rect

pub fn set_origin(&mut self, x: i32, y: i32)

Set the origin coordinates of the printable rect

pub fn scale(&mut self, scale_x: f32, scale_y: f32)

Scale the printable rect

pub fn rotate(&mut self, angle: f32)

Rotate the printable rect

pub fn translate(&mut self, x: i32, y: i32)

Translate the printable rect

pub fn untranslate(&mut self)

Untranslate the printable rect

pub fn is_current(&self) -> bool

Check whether the printer is the current printer

pub fn set_current(&mut self)

Set the printer to be the current printer

pub fn print_widget<W: WidgetExt>(&self, widget: &W, delta_x: i32, delta_y: i32)

Print a widget

pub fn print_window<W: WindowExt>(&self, win: &W, x_offset: i32, y_offset: i32)

Print a window

pub fn set_dialog_title(msg: &'static str)

Set the dialog “Title”

pub fn set_dialog_printer(msg: &'static str)

Set the dialog “Printer”

pub fn set_dialog_range(msg: &'static str)

Set dialog “Range”

pub fn set_dialog_copies(msg: &'static str)

Set dialog “Copies”

pub fn set_dialog_all(msg: &'static str)

Set dialog “All”

pub fn set_dialog_pages(msg: &'static str)

Set dialog “Pages”

pub fn set_dialog_from(msg: &'static str)

Set dialog “From”

pub fn set_dialog_to(msg: &'static str)

Set dialog “To”

pub fn set_dialog_properties(msg: &'static str)

Set dialog “Properties”

pub fn set_dialog_copy_number(msg: &'static str)

Set dialog “Number of copies”

pub fn set_dialog_print_button(msg: &'static str)

Set dialog “Print” button

pub fn set_dialog_cancel_button(msg: &'static str)

Set dialog “Cancel” button

pub fn set_dialog_print_to_file(msg: &'static str)

Set dialog “Print to file” button

pub fn set_property_title(msg: &'static str)

Set property “Title”

pub fn set_property_pagesize(msg: &'static str)

Set property “Page Size”

pub fn set_property_mode(msg: &'static str)

Set property “Mode”

pub fn set_property_use(msg: &'static str)

Set property “Use”

pub fn set_property_save(msg: &'static str)

Set property “Save”

pub fn set_property_cancel(msg: &'static str)

Set property “Cancel”