fltk::text

Struct TextBuffer

pub struct TextBuffer { /* private fields */ }

Wraps a text buffer, Cloning a text buffer invalidates the underlying pointer, thus the no derive(Clone)

Implementations

impl TextBuffer

pub unsafe fn delete(buf: Self)

Deletes the TextBuffer

Safety

The buffer shouldn’t be deleted while the Display widget still needs it

pub unsafe fn delete_buffer(buf: Self)

Deletes the TextBuffer

Safety

The buffer shouldn’t be deleted while the Display widget still needs it

pub unsafe fn from_ptr(ptr: *mut Fl_Text_Buffer) -> Self

Initializes a text buffer from a pointer

Safety

The pointer must be valid

pub unsafe fn as_ptr(&self) -> *mut Fl_Text_Buffer

Returns the inner pointer from a text buffer

Safety

Can return multiple mutable pointers to the same buffer

pub fn set_text(&mut self, txt: &str)

Sets the text of the buffer

pub fn text(&self) -> String

Returns the text of the buffer

pub fn append(&mut self, text: &str)

Appends to the buffer. To append and scroll to the end of the buffer:

use fltk::{prelude::*, *};
let txt = "Some long text!";
let buf = text::TextBuffer::default();
let mut disp = text::TextDisplay::default();
disp.set_buffer(Some(buf));
disp.buffer().unwrap().append(txt);
disp.set_insert_position(disp.buffer().unwrap().length());
disp.scroll(
    disp.count_lines(0, disp.buffer().unwrap().length(), true),
    0,
);

pub fn append2(&mut self, text: &[u8])

Append bytes to the buffer

pub fn length(&self) -> i32

Get the length of the buffer

pub fn remove(&mut self, start: i32, end: i32)

Removes from the buffer

pub fn text_range(&self, start: i32, end: i32) -> Option<String>

Returns the text within the range

pub fn insert(&mut self, pos: i32, text: &str)

Inserts text into a position

pub fn replace(&mut self, start: i32, end: i32, text: &str)

Replaces text from position start to end

pub fn copy_from( &mut self, source_buf: &TextBuffer, start: i32, end: i32, to: i32, )

Copies text from a source buffer into the current buffer

pub fn copy(&self) -> TextBuffer

Copies whole text from a source buffer into a new buffer

pub fn undo(&mut self) -> Result<(), FltkError>

Performs an undo operation on the buffer

Errors

Errors on failure to undo

pub fn redo(&mut self) -> Result<i32, FltkError>

Performs a redo operation on the buffer. Returns the cursor position.

Errors

Errors on failure to undo

pub fn can_undo(&mut self, flag: bool)

Sets whether the buffer can undo

pub fn get_can_undo(&mut self) -> bool

Gets whether the buffer can undo

pub fn can_redo(&mut self) -> bool

Gets whether the buffer can redo

pub fn load_file<P: AsRef<Path>>(&mut self, path: P) -> Result<(), FltkError>

Loads a file into the buffer

Errors

Errors on failure to load file

pub fn save_file<P: AsRef<Path>>(&mut self, path: P) -> Result<(), FltkError>

Saves a buffer into a file

Errors

Errors on failure to save file

pub fn tab_distance(&self) -> i32

Returns the tab distance for the buffer

pub fn set_tab_distance(&mut self, tab_dist: i32)

Sets the tab distance

pub fn select(&mut self, start: i32, end: i32)

Selects the text from start to end

pub fn selected(&self) -> bool

Returns whether text is selected

pub fn unselect(&mut self)

Unselects text

pub fn selection_position(&self) -> Option<(i32, i32)>

Returns the selection position

pub fn selection_text(&self) -> String

Returns the selection text

pub fn remove_selection(&mut self)

Removes the selection

pub fn replace_selection(&mut self, text: &str)

Replaces selection

pub fn secondary_select(&mut self, start: i32, end: i32)

Secondary selects the text from start to end

pub fn secondary_selected(&self) -> bool

Returns whether text is secondary selected

pub fn secondary_unselect(&mut self)

Unselects text (secondary selection)

pub fn secondary_selection_position(&self) -> Option<(i32, i32)>

Returns the secondary selection position

pub fn secondary_selection_text(&self) -> String

Returns the secondary selection text

pub fn remove_secondary_selection(&mut self)

Removes the secondary selection

pub fn replace_secondary_selection(&mut self, text: &str)

Replaces the secondary selection

pub fn highlight(&mut self, start: i32, end: i32)

Highlights selection

pub fn is_highlighted(&self) -> bool

Returns whether text is highlighted

pub fn unhighlight(&mut self)

Unhighlights text

pub fn highlight_position(&self) -> Option<(i32, i32)>

Returns the highlight position

pub fn highlight_text(&self) -> String

Returns the highlighted text

pub fn line_text(&self, pos: i32) -> String

Returns the line at pos

pub fn line_start(&self, pos: i32) -> i32

Returns the index of the line’s start position at pos

pub fn word_start(&self, pos: i32) -> i32

Returns the index of the first character of a word at pos

pub fn word_end(&self, pos: i32) -> i32

Returns the index of the last character of a word at pos

pub fn count_lines(&self, start: i32, end: i32) -> i32

Counts the lines from start to end

pub fn call_modify_callbacks(&mut self)

Calls the modify callbacks

pub fn add_modify_callback<F: FnMut(i32, i32, i32, i32, &str) + 'static>( &mut self, cb: F, )

Adds a modify callback. callback args: pos: i32, inserted items: i32, deleted items: i32, restyled items: i32, deleted_text

pub fn remove_modify_callback<F: FnMut(i32, i32, i32, i32, &str) + 'static>( &mut self, cb: F, )

Removes a modify callback. callback args: pos: i32, inserted items: i32, deleted items: i32, restyled items: i32, deleted_text

pub fn search_forward( &self, start_pos: i32, search_string: &str, match_case: bool, ) -> Option<i32>

Forward search for a string

pub fn search_backward( &self, start_pos: i32, search_string: &str, match_case: bool, ) -> Option<i32>

Backward search for a string

pub fn find_char_forward( &self, start_pos: i32, search_char: char, ) -> Option<i32>

Forward search for a char

pub fn find_char_backward( &self, start_pos: i32, search_char: char, ) -> Option<i32>

Backward search for a char