fltk

Module valuator

Valuator widgets

Structs

Adjuster

Creates an adjuster widget

Counter

Creates a counter widget

Dial

Creates a dial widget

FillDial

Creates a fill dial

FillSlider

Creates a fill slider

HorFillSlider

Creates a horizontal fill slider

HorNiceSlider

Creates a horizontal nice slider

HorSlider

Creates a horizontal slider

HorValueSlider

Creates a horizontal value slider

LineDial

Creates a line dial widget

NiceSlider

Creates a nice slider widget

Roller

Creates a roller widget

Scrollbar

Creates a scrollbar widget

Slider

Creates a slider widget

ValueInput

Creates an value input widget, which takes a numeric value. If a step is set, the value can be also dragged

ValueOutput

Creates an value output widget

ValueSlider

Creates a value slider widget

Enums

CounterType

Defines counter types

DialType

Defines dial types

ScrollbarType

Defines scrollbar types

SliderType

Defines slider types